+++ scaler/io/ymq/common.h
@@ -14,7 +14,7 @@ const size_t HEADER_SIZE = 4;  // size of the message header in bytes
 
 using Errno = int;
 
-void print_trace(void) {
+inline void print_trace(void) {
     void* array[10];
     char** strings;
     int size, i;
@@ -32,7 +32,7 @@ void print_trace(void) {
 
 // this is an unrecoverable error that exits the program
 // prints a message plus the source location
-[[noreturn]] void panic(std::string message, const std::source_location& location = std::source_location::current()) {
+[[noreturn]] inline void panic(std::string message, const std::source_location& location = std::source_location::current()) {
     auto file_name = std::string(location.file_name());
     file_name      = file_name.substr(file_name.find_last_of("/") + 1);
 
@@ -44,7 +44,7 @@ void print_trace(void) {
     std::abort();
 }
 
-[[noreturn]] void todo(
+[[noreturn]] inline void todo(
     std::optional<std::string> message   = std::nullopt,
     const std::source_location& location = std::source_location::current()) {
     if (message) {
@@ -54,19 +54,19 @@ void print_trace(void) {
     }
 }
 
-uint8_t* datadup(const uint8_t* data, size_t len) {
+inline uint8_t* datadup(const uint8_t* data, size_t len) {
     uint8_t* dup = new uint8_t[len];
     std::memcpy(dup, data, len);
     return dup;
 }
 
-void serialize_u32(uint32_t x, uint8_t buffer[4]) {
+inline void serialize_u32(uint32_t x, uint8_t buffer[4]) {
     buffer[0] = x & 0xFF;
     buffer[1] = (x >> 8) & 0xFF;
     buffer[2] = (x >> 16) & 0xFF;
     buffer[3] = (x >> 24) & 0xFF;
 }
 
-void deserialize_u32(const uint8_t buffer[4], uint32_t* x) {
+inline void deserialize_u32(const uint8_t buffer[4], uint32_t* x) {
     *x = buffer[0] | buffer[1] << 8 | buffer[2] << 16 | buffer[3] << 24;
 }
