--- scaler/io/ymq/defs.h
@@ -1,1 +1,0 @@
-// TODO: C-API definitions
