+++ scaler/io/ymq/event_loop_thread.cpp
@@ -1,10 +1,11 @@
 
 #include "scaler/io/ymq/event_loop_thread.h"
 
-// IOSocket* EventLoopThread::addIOSocket(std::string identity, std::string socketType) {
-//     return nullptr;
-// }
+void EventLoopThread::addIOSocket(std::shared_ptr<IOSocket>) {
+    // todo
+}
 
-// bool EventLoopThread::removeIOSocket(IOSocket*) {
-//     return false;
-// }
+bool EventLoopThread::removeIOSocket(std::shared_ptr<IOSocket>) {
+    // todo
+    return false;
+}
