+++ scaler/io/ymq/examples/CMakeLists.txt
@@ -1,4 +1,4 @@
 
 add_executable(echo_server echo_server.cpp)
 
-target_link_libraries(echo_server ymq)
+target_link_libraries(echo_server cc_ymq)
