+++ scaler/io/ymq/interruptive_concurrent_queue.h
@@ -5,7 +5,7 @@
 #include <optional>
 
 // First-party
-#include "common.h"
+// #include "common.h"
 #include "file_descriptor.h"
 
 // Third-party
