+++ scaler/io/ymq/pymod_ymq/io_socket.h
@@ -1,14 +1,13 @@
-// allows us to define PyTypeObjects in the canonical way without warnings
-#pragma clang diagnostic ignored "-Wreorder-init-list"
-#pragma clang diagnostic ignored "-Wc99-designator"
-
 // Python
 #define PY_SSIZE_T_CLEAN
 #include <Python.h>
 #include <structmember.h>
 
+struct IOSocket;
+
 struct PyIOSocket {
     PyObject_HEAD;
+    IOSocket* socket;  // Use shared_ptr for memory management
 };
 
 static int PyIOSocket_init(PyIOSocket* self, PyObject* args, PyObject* kwds) {
@@ -18,6 +17,8 @@ static int PyIOSocket_init(PyIOSocket* self, PyObject* args, PyObject* kwds) {
 static void PyIOSocket_dealloc(PyIOSocket* self) {
     // todo
 }
+
+// static PyObject* PyIOSocket_send(PyIOSocket* self, PyObject* args) {}
 
 static PyObject* PyIOSocket_repr(PyIOSocket* self) {
     Py_RETURN_NONE;  // todo
@@ -33,17 +34,17 @@ static PyMethodDef PyIOSocket_methods[] = {{nullptr, nullptr, 0, nullptr}};
 
 // clang-format off
 static PyTypeObject PyIOSocketType = {
-    PyVarObject_HEAD_INIT(NULL, 0)
+    .ob_base = PyVarObject_HEAD_INIT(NULL, 0)
     .tp_name      = "ymq.IOSocket",
-    .tp_doc       = PyDoc_STR("IOSocket"),
     .tp_basicsize = sizeof(PyIOSocket),
     .tp_itemsize  = 0,
-    .tp_flags     = Py_TPFLAGS_DEFAULT,
-    .tp_new       = PyType_GenericNew,
-    .tp_init      = (initproc)PyIOSocket_init,
-    .tp_repr      = (reprfunc)PyIOSocket_repr,
     .tp_dealloc   = (destructor)PyIOSocket_dealloc,
-    .tp_getset    = PyIOSocket_properties,
+    .tp_repr      = (reprfunc)PyIOSocket_repr,
+    .tp_flags     = Py_TPFLAGS_DEFAULT,
+    .tp_doc       = PyDoc_STR("IOSocket"),
     .tp_methods   = PyIOSocket_methods,
+    .tp_getset    = PyIOSocket_properties,
+    .tp_init      = (initproc)PyIOSocket_init,
+    .tp_new       = PyType_GenericNew,
 };
 // clang-format on
